add stacks.html
