+++ card_app/templates/cards.html
@@ -27,6 +27,7 @@
         {% endfor %}
     </table>
     <a href="/">Home</a>
+    <a href='/stacks'>View Stacks</a>
 </body>
 
 </html>
