+++ card_app/templates/index.html
@@ -11,6 +11,7 @@
 <body>
     <h1>CARD APP TESTING</h1>
     <a href='/cards'>View Cards</a>
+    <a href='/stacks'>View Stacks</a>
 </body>
 
 </html>
