+++ server.py
@@ -1,5 +1,5 @@
 from card_app import app
-from card_app.controllers import cards
+from card_app.controllers import cards, stacks
 
 if __name__ == '__main__':
     app.run(debug=True)
